ゲーム制作応用
マリオのがwalkからidleに遷移するときの動きをプログラムで追加した

+++ Develop/Object/Player/PlayerState/WalkingState.h
@@ -6,7 +6,7 @@ class WalkingState : public PlayerStateBase
 {
 public:
 	InputManager* input;
-	double speed;
+	float speed;
 
 public:
 	WalkingState(class Player* p);
